requirements_local.txt
tensorflow>=2.15.0
pandas
numpy
scikit-learn
Pillow
matplotlib
seaborn
tqdm
